build: fix react-refresh warnings
