+++ src/components/ui/button-variants.ts
@@ -1,0 +1,33 @@
+import { cva } from 'class-variance-authority';
+
+const buttonVariants = cva(
+  'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
+  {
+    variants: {
+      variant: {
+        default:
+          'bg-primary text-primary-foreground hover:bg-primary/90 border border-primary',
+        destructive:
+          'bg-destructive text-destructive-foreground hover:bg-destructive/90 border border-destructive shadow-sm',
+        outline:
+          'border border-input bg-background hover:bg-accent hover:text-accent-foreground shadow-sm',
+        secondary:
+          'bg-secondary text-secondary-foreground hover:bg-secondary/80 border border-secondary shadow-sm',
+        ghost: 'hover:bg-accent hover:text-accent-foreground',
+        link: 'text-primary underline-offset-4 hover:underline',
+      },
+      size: {
+        default: 'h-10 px-4 py-2',
+        sm: 'h-9 rounded-md px-3',
+        lg: 'h-11 rounded-md px-8',
+        icon: 'h-10 w-10',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+      size: 'default',
+    },
+  },
+);
+
+export { buttonVariants };
