+++ src/components/ui/button.types.ts
@@ -1,0 +1,9 @@
+import type * as React from 'react';
+import type { VariantProps } from 'class-variance-authority';
+import { buttonVariants } from './button-variants';
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
+}
