+++ src/components/ui/pagination.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { ChevronLeft, ChevronRight, MoreHorizontal } from 'lucide-react';
 
 import { cn } from '@/lib/utils';
-import { ButtonProps, buttonVariants } from '@/components/ui/button';
+import { buttonVariants } from '@/components/ui/button-variants';
+import type { ButtonProps } from '@/components/ui/button.types';
 
 const Pagination = ({ className, ...props }: React.ComponentProps<'nav'>) => (
   <nav
